# api-spec/chatgpt-action-mcp-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Agent Factory PRD Submission via MCP Server",
    "description": "API for submitting Product Requirements Documents (PRDs) to the AI Agent Factory platform through the MCP server. ChatGPT can automatically detect when a conversation contains a PRD and submit it using this action. The MCP server handles delivery to the agent factory.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://ai-agent-factory-mcp-server-952475323593.us-central1.run.app",
      "description": "Production MCP HTTP Server"
    }
  ],
  "paths": {
    "/api/v1/prds/incoming": {
      "post": {
        "operationId": "submitPRD",
        "summary": "Submit a Product Requirements Document (PRD) via MCP Server",
        "description": "Automatically submit a PRD to the AI Agent Factory through the MCP server. ChatGPT should call this action when it detects that the user has created or described a PRD in the conversation. The PRD content should be formatted as markdown. The MCP server will handle delivery to the agent factory backend.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "description": "The complete PRD content in markdown format. Should include title, description, requirements, and other PRD sections.",
                    "example": "# PRD Title\n\n## Description\n\nA comprehensive description of the product or feature...\n\n## Requirements\n\n- Requirement 1\n- Requirement 2\n\n## Technical Requirements\n\n- Technical detail 1\n- Technical detail 2"
                  }
                },
                "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PRD submitted successfully via MCP server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Unique PRD identifier"
                    },
                    "title": {
                      "type": "string",
                      "description": "PRD title"
                    },
                    "status": {
                      "type": "string",
                      "description": "PRD status"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Creation timestamp"
                    },
                    "prd_id": {
                      "type": "string",
                      "description": "PRD ID (from MCP server response)"
                    },
                    "message": {
                      "type": "string",
                      "description": "Success message from MCP server"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid PRD content or format"
          },
          "500": {
            "description": "Server error (MCP server or backend)"
          }
        }
      }
    }
  }
}